Reorganized scss files
