+++ src/scss/_npc-card.scss
@@ -345,15 +345,3 @@
 		}
 	}
 }
-
-.active {
-	opacity: 1;
-}
-
-.hidden {
-	opacity: 0;
-}
-
-.gone {
-	display: none !important;
-}
