+++ src/scss/layouts/_layouts.scss
@@ -1,0 +1,3 @@
+@import "./app-header";
+@import "./app-footer";
+@import "./pre-footer";
